+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/member/infrastructure/persistence/jpa/MemberJpaRepository.java
@@ -1,15 +1,11 @@
-package com.nextdoor.nextdoor.domain.member.repository;
+package com.nextdoor.nextdoor.domain.member.infrastructure.persistence.jpa;
 
 import com.nextdoor.nextdoor.domain.member.domain.model.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MemberRepository {
-
-    Member save(Member member);
-
-    Optional<Member> findById(Long id);
+public interface MemberJpaRepository extends JpaRepository<Member, Long> {
 
     Optional<Member> findByUserKey(String userKey);
 }
